+++ src/modules/navigation/components/AppNavigator.tsx
@@ -3,6 +3,7 @@ import { AppStackParamList } from '../../../types/navigation';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
 import { mainTheme } from '../../../themes/mainTheme';
 import RootScreen from '../../root/RootScreen';
+import ChatMessagesScreen from '../../chat/messages/ChatMessagesScreen';
 
 const Stack = createNativeStackNavigator<AppStackParamList>();
 
@@ -12,6 +13,9 @@ const AppNavigator = () => {
       headerTintColor: mainTheme.PRIMARY_COLOR
     }}>
         <Stack.Screen name="Root" component={RootScreen} options={{
+          headerShown: false,
+        }} />
+        <Stack.Screen name="ChatMessages" component={ChatMessagesScreen} options={{
           headerShown: false,
         }} />
     </Stack.Navigator>
