+++ src/modules/navigation/hooks/useAppNavigation.tsx
@@ -1,0 +1,9 @@
+import { NavigationProp, useNavigation } from '@react-navigation/native'
+import { AppStackParamList } from '../../../types/navigation'
+
+const useAppNavigation = () => {
+    const navigation = useNavigation<NavigationProp<AppStackParamList>>();
+    return navigation;
+}
+
+export default useAppNavigation;
