+++ src/modules/navigation/hooks/useRootNavigation.tsx
@@ -1,0 +1,9 @@
+import { NavigationProp, useNavigation } from '@react-navigation/native'
+import { RootScreenTabParamList } from '../../../types/navigation'
+
+const useRootNavigation = () => {
+    const navigation = useNavigation<NavigationProp<RootScreenTabParamList>>();
+    return navigation;
+}
+
+export default useRootNavigation;
